Changed the package from mllib to ml and indentation to 2 spaces for all the Java code
