+++ examples/src/main/scala/org/apache/spark/examples/ml/EstimatorTransformerParamExample.scala
@@ -16,7 +16,7 @@
  */
 
 // scalastyle:off println
-package org.apache.spark.examples.mllib
+package org.apache.spark.examples.ml
 
 import org.apache.spark.{ SparkConf, SparkContext }
 import org.apache.spark.ml.classification.LogisticRegression
