+++ examples/src/main/scala/org/apache/spark/examples/ml/ModelSelectionViaCrossValidationExample.scala
@@ -16,56 +16,74 @@
  */
 
 // scalastyle:off println
-package org.apache.spark.examples.mllib
+package org.apache.spark.examples.ml
 
 import org.apache.spark.{ SparkConf, SparkContext }
-import org.apache.spark.ml.{ Pipeline, PipelineModel }
+import org.apache.spark.ml.Pipeline
 import org.apache.spark.ml.classification.LogisticRegression
+import org.apache.spark.ml.evaluation.BinaryClassificationEvaluator
 import org.apache.spark.ml.feature.{ HashingTF, Tokenizer }
+import org.apache.spark.ml.tuning.{ CrossValidator, ParamGridBuilder }
 import org.apache.spark.mllib.linalg.Vector
 import org.apache.spark.sql.Row
 import org.apache.spark.sql.SQLContext
 
-object PipelineExample {
+object ModelSelectionViaCrossValidationExample {
 
   def main(args: Array[String]): Unit = {
-    val conf = new SparkConf().setAppName("PipelineExample")
+    val conf = new SparkConf().setAppName("ModelSelectionViaCrossValidationExample")
     val sc = new SparkContext(conf)
     val sqlContext = new SQLContext(sc)
 
     // $example on$
-    // Prepare training documents from a list of (id, text, label) tuples.
+    // Prepare training data from a list of (id, text, label) tuples.
     val training = sqlContext.createDataFrame(Seq(
       (0L, "a b c d e spark", 1.0),
       (1L, "b d", 0.0),
       (2L, "spark f g h", 1.0),
-      (3L, "hadoop mapreduce", 0.0))).toDF("id", "text", "label")
+      (3L, "hadoop mapreduce", 0.0),
+      (4L, "b spark who", 1.0),
+      (5L, "g d a y", 0.0),
+      (6L, "spark fly", 1.0),
+      (7L, "was mapreduce", 0.0),
+      (8L, "e spark program", 1.0),
+      (9L, "a e c l", 0.0),
+      (10L, "spark compile", 1.0),
+      (11L, "hadoop software", 0.0))).toDF("id", "text", "label")
 
     // Configure an ML pipeline, which consists of three stages: tokenizer, hashingTF, and lr.
     val tokenizer = new Tokenizer()
       .setInputCol("text")
       .setOutputCol("words")
     val hashingTF = new HashingTF()
-      .setNumFeatures(1000)
       .setInputCol(tokenizer.getOutputCol)
       .setOutputCol("features")
     val lr = new LogisticRegression()
       .setMaxIter(10)
-      .setRegParam(0.01)
     val pipeline = new Pipeline()
       .setStages(Array(tokenizer, hashingTF, lr))
 
-    // Fit the pipeline to training documents.
-    val model = pipeline.fit(training)
+    // We use a ParamGridBuilder to construct a grid of parameters to search over.
+    // With 3 values for hashingTF.numFeatures and 2 values for lr.regParam,
+    // this grid will have 3 x 2 = 6 parameter settings for CrossValidator to choose from.
+    val paramGrid = new ParamGridBuilder()
+      .addGrid(hashingTF.numFeatures, Array(10, 100, 1000))
+      .addGrid(lr.regParam, Array(0.1, 0.01))
+      .build()
 
-    // now we can optionally save the fitted pipeline to disk
-    model.save("/tmp/spark-logistic-regression-model")
+    // We now treat the Pipeline as an Estimator, wrapping it in a CrossValidator instance.
+    // This will allow us to jointly choose parameters for all Pipeline stages.
+    // A CrossValidator requires an Estimator, a set of Estimator ParamMaps, and an Evaluator.
+    // Note that the evaluator here is a BinaryClassificationEvaluator and its default metric
+    // is areaUnderROC.
+    val cv = new CrossValidator()
+      .setEstimator(pipeline)
+      .setEvaluator(new BinaryClassificationEvaluator)
+      .setEstimatorParamMaps(paramGrid)
+      .setNumFolds(2) // Use 3+ in practice
 
-    // we can also save this unfit pipeline to disk
-    pipeline.save("/tmp/unfit-lr-model")
-
-    // and load it back in during production
-    val sameModel = PipelineModel.load("/tmp/spark-logistic-regression-model")
+    // Run cross-validation, and choose the best set of parameters.
+    val cvModel = cv.fit(training)
 
     // Prepare test documents, which are unlabeled (id, text) tuples.
     val test = sqlContext.createDataFrame(Seq(
@@ -74,8 +92,8 @@ object PipelineExample {
       (6L, "mapreduce spark"),
       (7L, "apache hadoop"))).toDF("id", "text")
 
-    // Make predictions on test documents.
-    model.transform(test)
+    // Make predictions on test documents. cvModel uses the best model found (lrModel).
+    cvModel.transform(test)
       .select("id", "text", "probability", "prediction")
       .collect()
       .foreach {
